[REF] New individual/ directory

+++ examples/example1.cc
@@ -14,13 +14,13 @@
 #include <iostream>
 
 #include "kernel/environment.h"
-#include "kernel/individual.h"
+#include "kernel/individual/mep.h"
 #include "kernel/src/primitive/factory.h"
 
 int main(int argc, char *argv[])
 {
   vita::environment env(true);
-  env.code_length = argc > 1 ? atoi(argv[1]) : 10;
+  env.code_length = static_cast<unsigned>(argc > 1 ? std::atoi(argv[1]) : 10);
 
   vita::symbol_set sset;
